// src/types/movieType.ts
interface MoviesType {
	adult: boolean
	backdrop_path: string
	genre_ids: number[]
	id: number
	original_language: string
	orginal_title: string
	overview: string
	popularity: number
	poster_path: string
	release_date: string
	title: string
	video: boolean
	vote_average: number
	vote_count: number
	price: number
}

export default MoviesType;